Converted explicit Display setters to lombok chaining setters
Moved DisplayLine into it's own class file
